chore: Move reaction bottom sheet to chat package
