+++ app/src/main/java/com/nextcloud/talk/chat/ui/ShowReactionsSheet.kt
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-package com.nextcloud.talk.ui.dialog
+package com.nextcloud.talk.chat.ui
 
 import android.content.res.Configuration
 import android.util.Log
@@ -25,9 +25,9 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.PrimaryScrollableTabRow
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Tab
 import androidx.compose.material3.Text
-import androidx.compose.material3.Surface
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.material3.rememberModalBottomSheetState
@@ -164,14 +164,15 @@ internal fun ReactionsSheetLayout(
     val tabs: List<String?> = listOf(null) + reactions.keys.toList()
 
     Column(
-        modifier = Modifier
+        modifier = Modifier.Companion
             .fillMaxWidth()
             .padding(bottom = 24.dp)
     ) {
         val reactionsTotal = reactions.values.sum()
         PrimaryScrollableTabRow(
             selectedTabIndex = selectedTabIndex,
-            containerColor = MaterialTheme.colorScheme.surfaceContainerLow
+            containerColor = MaterialTheme.colorScheme.surfaceContainerLow,
+            edgePadding = 0.dp
         ) {
             tabs.forEachIndexed { index, emoji ->
                 val count = if (emoji == null) reactionsTotal else reactions[emoji] ?: 0
@@ -188,7 +189,7 @@ internal fun ReactionsSheetLayout(
             }
         }
         LazyColumn(
-            modifier = Modifier
+            modifier = Modifier.Companion
                 .fillMaxWidth()
                 .heightIn(max = 288.dp)
         ) {
@@ -229,9 +230,11 @@ private fun ReactionVoterRow(
                     ?: guestLabel
                 ApiUtils.getUrlForGuestAvatar(user.baseUrl, displayName, false)
             }
+
             ReactionVoter.ReactionActorType.USERS -> {
                 ApiUtils.getUrlForAvatar(user.baseUrl, reactionItem.reactionVoter.actorId, false, isDark)
             }
+
             else -> null
         }
     }
@@ -247,31 +250,31 @@ private fun ReactionVoterRow(
     }
 
     Row(
-        modifier = Modifier
+        modifier = Modifier.Companion
             .fillMaxWidth()
             .clickable(enabled = canDelete && reactionItem.reaction != null) {
                 reactionItem.reaction?.let(onDeleteReaction)
             }
             .padding(horizontal = 16.dp, vertical = 8.dp),
-        verticalAlignment = Alignment.CenterVertically
+        verticalAlignment = Alignment.Companion.CenterVertically
     ) {
         AsyncImage(
             model = avatarRequest ?: R.drawable.account_circle_96dp,
             contentDescription = null,
             placeholder = painterResource(R.drawable.account_circle_96dp),
             error = painterResource(R.drawable.account_circle_96dp),
-            modifier = Modifier.size(40.dp)
+            modifier = Modifier.Companion.size(40.dp)
         )
-        Spacer(modifier = Modifier.width(16.dp))
+        Spacer(modifier = Modifier.Companion.width(16.dp))
         Text(
             text = reactionItem.reactionVoter.actorDisplayName ?: "",
             style = MaterialTheme.typography.bodyLarge,
-            modifier = Modifier.weight(1f)
+            modifier = Modifier.Companion.weight(1f)
         )
         Text(
             text = reactionItem.reaction ?: "",
             style = MaterialTheme.typography.titleMedium,
-            modifier = Modifier.padding(start = 8.dp)
+            modifier = Modifier.Companion.padding(start = 8.dp)
         )
     }
 }
